+++ librepair/add_missing.c
@@ -83,8 +83,9 @@ static errno_t repair_am_node_prepare(repair_am_t *am, reiser4_node_t *node) {
 	for (place.pos.item = 0; place.pos.item < count; place.pos.item++) {
 		if ((res = reiser4_place_fetch(&place))) {
 			aal_error("Node (%llu), item (%u): failed to open the "
-				  "item.",node->block->nr, place.pos.item);
-
+				  "item.",
+				  (unsigned long long)node->block->nr,
+				  place.pos.item);
 			return res;
 		}
 
@@ -155,7 +156,8 @@ static errno_t repair_am_nodes_insert(repair_am_t *am,
 
 		if (node == NULL) {
 			aal_error("Add Missing pass failed to "
-				  "open the node (%llu)", blk);
+				  "open the node (%llu)",
+				  (unsigned long long)blk);
 
 			return -EINVAL;
 		}
@@ -177,7 +179,8 @@ static errno_t repair_am_nodes_insert(repair_am_t *am,
 		
 		if (res < 0 && res != -ESTRUCT) {
 			aal_error("Add missing pass failed to attach "
-				  "the node (%llu) to the tree.", blk);
+				  "the node (%llu) to the tree.",
+				  (unsigned long long)blk);
 
 			goto error_close_node;
 		} else if (res == 0) {
@@ -235,7 +238,8 @@ static errno_t repair_am_items_insert(repair_am_t *am,
 		
 		if (node == NULL) {
 			aal_error("Add Missing pass failed to "
-				  "open the node (%llu)", blk);
+				  "open the node (%llu)",
+				  (unsigned long long)blk);
 			return -EINVAL;
 		}
 
@@ -252,7 +256,9 @@ static errno_t repair_am_items_insert(repair_am_t *am,
 			if ((res = reiser4_place_fetch(&place))) {
 				aal_error("Node (%llu), item (%u): "
 					  "cannot open the item "
-					  "place.", blk, pos->item);
+					  "place.",
+					  (unsigned long long)blk,
+					  pos->item);
 
 				goto error_close_node;
 			}
