+++ plugin/sdext/sdext_lw/sdext_lw_repair.c
@@ -18,7 +18,8 @@ errno_t sdext_lw_check_struct(stat_entity_t *stat, repair_hint_t *hint) {
 	if (stat->offset + sizeof(sdext_lw_t) > stat->place->len) {
 		fsck_mess("Node (%llu), item (%u), [%s]: does not look "
 			  "like a valid (%s) statdata extension.", 
-			  place_blknr(stat->place), stat->place->pos.item,
+			  (unsigned long long)place_blknr(stat->place),
+			  stat->place->pos.item,
 			  print_key(sdext_lw_core, &stat->place->key),
 			  stat->plug->p.label);
 		return 	RE_FATAL;
